requirements.txt
asgiref==3.3.4
black==22.1.0
click==8.0.3
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.3
django-crispy-forms==1.7.0
django-extensions==3.1.3
django-grappelli==3.0.2
django-heroku==0.3.1
djangorestframework==3.13.1
dr-scaffold==2.1.2
gunicorn==20.1.0
inflect==5.4.0
isort==5.10.1
Markdown==3.3.6
mypy-extensions==0.4.3
mysql-connector-python==8.0.25
pathspec==0.9.0
platformdirs==2.5.0
protobuf==3.19.4
psycopg2==2.9.3
pytz==2021.1
sqlparse==0.4.1
tomli==2.0.1
whitenoise==6.0.0
